fixed getError function, additional cast was added to ensure the result has the right data type

+++ ax_blackscholes/bellerophon/bellerophon_functions.cpp
@@ -56,7 +56,7 @@ extern "C" double BELLERO_getError() {
 		count++;
 	}
 
-	avgError = absError/count;
+	avgError = (double)absError/count;
 
 	return avgError;
 }
